+++ apps/bfd-model/bfd-model-rda/src/main/java/gov/cms/bfd/model/rda/PreAdjMcsClaim.java
@@ -143,7 +143,7 @@ public class PreAdjMcsClaim {
   private String idrClmLevelInd;
 
   @Column(name = "`idrHdrAudit`")
-  private Short idrHdrAudit;
+  private Integer idrHdrAudit;
 
   @Column(name = "`idrHdrAuditInd`", length = 1)
   private String idrHdrAuditInd;
